Take two.
