+++ sandbox/src/Identity.jsx
@@ -13,8 +13,7 @@
 /* eslint-disable no-console, func-names */
 
 import React, { useEffect, useState } from "react";
-import setupAlloy from "./helpers/setupAlloy";
-import configureAlloy from "./helpers/configureAlloy";
+import useAlloy from "./helpers/useAlloy";
 
 const readCookies = () => {
   const cookies = {};
@@ -99,10 +98,9 @@ export default function Identity() {
   const [currentIdentityCookie, setCurrentIdentityCookie] = useState("");
   const [identity, setIdentity] = useState("");
 
-  useEffect(() => {
-    setupAlloy();
-    configureAlloy();
+  useAlloy();
 
+  useEffect(() => {
     const ecid = readIdentityCookie();
     setOriginalIdentityCookie(ecid);
     setCurrentIdentityCookie(ecid);
