+++ sandbox/src/Links.jsx
@@ -1,8 +1,7 @@
-import React, { useEffect } from "react";
+import React from "react";
 import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import sendPageViewEvent from "./helpers/sendPageViewEvent";
-import configureAlloy from "./helpers/configureAlloy";
-import setupAlloy from "./helpers/setupAlloy";
+import useAlloy from "./helpers/useAlloy";
+import useSendPageViewEvent from "./helpers/useSendPageViewEvent";
 
 const adobeLink = () => {
   window.alloy("sendEvent", {
@@ -14,11 +13,8 @@ const adobeLink = () => {
 };
 
 export default function Links() {
-  useEffect(() => {
-    setupAlloy();
-    configureAlloy();
-    sendPageViewEvent();
-  }, []);
+  useAlloy();
+  useSendPageViewEvent();
 
   return (
     <div>
