+++ sandbox/src/Personalization.jsx
@@ -1,15 +1,11 @@
-import React, { useEffect } from "react";
+import React from "react";
 import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import configureAlloy from "./helpers/configureAlloy";
-import sendPageViewEvent from "./helpers/sendPageViewEvent";
-import setupAlloy from "./helpers/setupAlloy";
+import useAlloy from "./helpers/useAlloy";
+import useSendPageViewEvent from "./helpers/useSendPageViewEvent";
 
 export default function Personalization() {
-  useEffect(() => {
-    setupAlloy();
-    configureAlloy();
-    sendPageViewEvent({ renderDecisions: true });
-  }, []);
+  useAlloy();
+  useSendPageViewEvent({ renderDecisions: true });
 
   return (
     <div>
