+++ sandbox/src/PersonalizationFormBased.jsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useState } from "react";
 import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import sendPageViewEvent from "./helpers/sendPageViewEvent";
-import configureAlloy from "./helpers/configureAlloy";
-import setupAlloy from "./helpers/setupAlloy";
+import useAlloy from "./helpers/useAlloy";
+import useSendPageViewEvent from "./helpers/useSendPageViewEvent";
 
 const SCOPES_FOR_PAGE = ["sandbox-personalization-page2"];
 
@@ -17,15 +16,13 @@ export default function Personalization() {
   const [renderCounter, setRenderCounter] = useState(0);
   const [propositions, setPropositions] = useState(undefined);
 
-  useEffect(() => {
-    setupAlloy();
-    configureAlloy();
-    sendPageViewEvent({
-      renderDecisions: true,
-      setPropositions,
-      decisionScopes: SCOPES_FOR_PAGE, // Note: this option will soon be deprecated, please use personalization.decisionScopes instead
-    });
-  }, []);
+  useAlloy();
+
+  useSendPageViewEvent({
+    renderDecisions: true,
+    setPropositions,
+    decisionScopes: SCOPES_FOR_PAGE, // Note: this option will soon be deprecated, please use personalization.decisionScopes instead
+  });
 
   useEffect(() => {
     if (propositions) {
