+++ sandbox/src/PersonalizationProfile.jsx
@@ -1,24 +1,21 @@
-import React, { useEffect } from "react";
+import React from "react";
 import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import sendPageViewEvent from "./helpers/sendPageViewEvent";
-import configureAlloy from "./helpers/configureAlloy";
-import setupAlloy from "./helpers/setupAlloy";
+import useAlloy from "./helpers/useAlloy";
+import useSendPageViewEvent from "./helpers/useSendPageViewEvent";
 
 export default function Personalization() {
-  useEffect(() => {
-    setupAlloy();
-    configureAlloy();
-    sendPageViewEvent({
-      renderDecisions: true,
-      data: {
-        __adobe: {
-          target: {
-            "profile.favoriteColor": "Black",
-          },
+  useAlloy();
+  useSendPageViewEvent({
+    renderDecisions: true,
+    data: {
+      __adobe: {
+        target: {
+          "profile.favoriteColor": "Black",
         },
       },
-    });
-  }, []);
+    },
+  });
+
   return (
     <div>
       <ContentSecurityPolicy />
