+++ sandbox/src/RedirectOffers.jsx
@@ -2,14 +2,11 @@
 
 import React, { useEffect } from "react";
 import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import setupAlloy from "./helpers/setupAlloy";
-import configureAlloy from "./helpers/configureAlloy";
+import useAlloy from "./helpers/useAlloy";
 
 export default function RedirectOffers() {
+  useAlloy();
   useEffect(() => {
-    setupAlloy();
-    configureAlloy();
-
     window
       .alloy("sendEvent", {
         renderDecisions: true,
